+++ sdk/clientcore/core/src/test/java/io/clientcore/core/http/pipeline/HttpPipelinePolicyTests.java
@@ -24,7 +24,7 @@ public class HttpPipelinePolicyTests {
         SyncPolicy policy2 = new SyncPolicy();
 
         HttpPipeline pipeline
-            = new HttpPipelineBuilder().httpClient(new NoOpHttpClient()).policies(policy1, policy2).build();
+            = new HttpPipelineBuilder().httpClient(new NoOpHttpClient()).addPolicy(policy1).addPolicy(policy2).build();
 
         pipeline.send(new HttpRequest(HttpMethod.GET, "http://localhost/")).close();
 
@@ -37,7 +37,7 @@ public class HttpPipelinePolicyTests {
         DefaultImplementationSyncPolicy policyWithDefaultSyncImplementation = new DefaultImplementationSyncPolicy();
 
         HttpPipeline pipeline = new HttpPipelineBuilder().httpClient(new NoOpHttpClient())
-            .policies(policyWithDefaultSyncImplementation)
+            .addPolicy(policyWithDefaultSyncImplementation)
             .build();
 
         pipeline.send(new HttpRequest(HttpMethod.GET, "http://localhost/")).close();
@@ -52,7 +52,7 @@ public class HttpPipelinePolicyTests {
     @Test
     public void doesNotThrowThatThreadIsNonBlocking() throws IOException {
         SyncPolicy policy1 = new SyncPolicy();
-        HttpPipelinePolicy badPolicy1 = (httpRequest, next) -> {
+        HttpPipelinePolicy badPolicy1 = (ignored, next) -> {
             try {
                 Thread.sleep(10);
             } catch (InterruptedException e) {
@@ -62,7 +62,7 @@ public class HttpPipelinePolicyTests {
             return next.process();
         };
 
-        HttpPipelinePolicy badPolicy2 = (httpRequest, next) -> {
+        HttpPipelinePolicy badPolicy2 = (ignored, next) -> {
             try {
                 Thread.sleep(10);
             } catch (InterruptedException e) {
@@ -81,8 +81,11 @@ public class HttpPipelinePolicyTests {
             return new HttpResponse<>(request, 200, new HttpHeaders(), null);
         };
 
-        HttpPipeline pipeline
-            = new HttpPipelineBuilder().httpClient(badClient).policies(policy1, badPolicy1, badPolicy2).build();
+        HttpPipeline pipeline = new HttpPipelineBuilder().httpClient(badClient)
+            .addPolicy(policy1)
+            .addPolicy(badPolicy1)
+            .addPolicy(badPolicy2)
+            .build();
 
         pipeline.send(new HttpRequest(HttpMethod.GET, "http://localhost/")).close();
     }