+++ sdk/cosmos/azure-cosmos-kafka-connect/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Release History
 
+### 2.1.1 (2025-02-08)
+
+#### Other Changes
+* Updated `azure-cosmos` to version `4.66.1`.
+
 ### 2.1.0 (2025-01-14)
 
 #### Bugs Fixed